LayoutTests/imported/w3c/web-platform-tests/html/semantics/popovers/popover-beforetoggle-opening-event.tentative-expected.txt


FAIL Ensure the `beforetoggle` event can be used to populate content before the popover renders assert_equals: expected "Show Event Occurred" but got ""
